Move CircuitPlayground specific component tests
