+++ apps/test/unit/lib/kits/maker/boards/circuitPlayground/NeoPixelTest.js
@@ -1,5 +1,5 @@
 /** @file Tests for our johnny-five Led.RGB wrapper */
-import {expect} from '../../../../util/deprecatedChai';
+import {expect} from '../../../../../../util/deprecatedChai';
 import sinon from 'sinon';
 import five from '@code-dot-org/johnny-five';
 import NeoPixel from '@cdo/apps/lib/kits/maker/boards/circuitPlayground/NeoPixel';
