+++ apps/test/unit/lib/kits/maker/boards/circuitPlayground/PiezoTest.js
@@ -1,5 +1,5 @@
 /** @file Tests for our johnny-five Piezo wrapper */
-import {expect} from '../../../../util/deprecatedChai';
+import {expect} from '../../../../../../util/deprecatedChai';
 import sinon from 'sinon';
 import five from '@code-dot-org/johnny-five';
 import Piezo from '@cdo/apps/lib/kits/maker/boards/circuitPlayground/Piezo';
