+++ apps/test/unit/lib/kits/maker/boards/circuitPlayground/SwitchTest.js
@@ -1,4 +1,4 @@
-import {expect} from '../../../../util/deprecatedChai';
+import {expect} from '../../../../../../util/deprecatedChai';
 import sinon from 'sinon';
 import {EventEmitter} from 'events'; // provided by webpack's node-libs-browser
 import Switch, {
